Update ping command to verify DB connection
The init-db-container.sh script includes a command to verify that the DB is ready before starting migrations. This command stopped working after upgrading the migration container from `Alpine 3.20` -> `Alpine 3.21` with the following error:

> error: 'TLS/SSL error: self-signed certificate in certificate chain'

Since this command is only verifying the connection is ready, skip SSL verification as a workaround. Also update the name of the command from `mysqladmin` -> `mariadb-admin` to fix a deprecation warning.